# docs/swagger.yaml
basePath: /
definitions:
...
  storage.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      id:
        type: integer
      isAdmin:
        type: boolean
      isDev:
        type: boolean
      name:
        type: string
      password:
        description: 已哈希的口令
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: lushi78778@xray.top
    name: GinkgoID Team
    url: https://xray.top
  description: 基于 Go(Gin) 的 OpenID Provider，实现 Discovery、JWKS、授权码+PKCE、Token、UserInfo、动态注册、注销、撤销、内省等接口。
  title: GinkgoID OIDC OP API
  version: 0.1.0
paths:
...
  /authorize:
    get:
      description: 实现授权码 + PKCE 及可选 Hybrid 流程
      parameters:
      - description: code 或 code id_token
        in: query
        name: response_type
        required: true
        type: string
      - description: 客户端 ID
        in: query
        name: client_id
        required: true
        type: string
      - description: 重定向 URI
        in: query
        name: redirect_uri
        required: true
        type: string
      - description: openid 等空格分隔
        in: query
        name: scope
        required: true
        type: string
      - description: 状态参数
        in: query
        name: state
        type: string
      - description: Hybrid 必需
        in: query
        name: nonce
        type: string
      - description: fragment|form_post|query
        in: query
        name: response_mode
        type: string
      - description: PKCE 挑战
        in: query
        name: code_challenge
        type: string
      - description: S256|plain
        in: query
        name: code_challenge_method
        type: string
      - description: none|login|consent
        in: query
        name: prompt
        type: string
      - description: ACR 要求
        in: query
        name: acr_values
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: 重定向到 redirect_uri
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 授权端点（Authorization）
      tags:
      - oauth2
...
  /consent:
    get:
      description: 显示授权范围供用户确认
      parameters:
      - description: 客户端 ID
        in: query
        name: client_id
        required: true
        type: string
      - description: 请求的 scope（空格分隔）
        in: query
        name: scope
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML
          schema:
            type: string
      summary: 授权同意页
      tags:
      - ui
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 处理用户对授权的同意或拒绝
      parameters:
      - description: approve 或 deny
        in: formData
        name: decision
        required: true
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: 重定向至 /authorize
          schema:
            type: string
      summary: 提交授权同意
      tags:
      - ui
  /dev/keys/rotate:
    post:
      description: （仅限非生产环境）触发一次新的 JWK 生成
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 开发辅助 - 轮换密钥
      tags:
      - dev
  /dev/users:
    get:
      description: （仅限非生产环境）列出所有用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.User'
            type: array
      summary: 开发辅助 - 用户列表
      tags:
      - dev
    post:
      consumes:
      - application/json
      description: （仅限非生产环境）快速创建用户
      parameters:
      - description: '{username,password}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 开发辅助 - 创建用户
      tags:
      - dev
...
  /login:
    get:
      description: 显示用户名密码登录表单
      produces:
      - text/html
      responses:
        "200":
          description: HTML
          schema:
            type: string
      summary: 登录页
      tags:
      - ui
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 处理用户名密码登录请求
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: 重定向至 /authorize
          schema:
            type: string
        "401":
          description: HTML 登录页（含错误）
          schema:
            type: string
      summary: 提交登录
      tags:
      - ui
...
  /userinfo:
    get:
      description: 使用 Access Token 获取用户 claims 信息
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        type: string
      - description: 当未设置 Authorization 时可使用
        in: query
        name: access_token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 用户信息端点（UserInfo）
      tags:
      - oidc
    post:
      description: 使用 Access Token 获取用户 claims 信息
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        type: string
      - description: 当未设置 Authorization 时可使用
        in: query
        name: access_token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 用户信息端点（UserInfo）
      tags:
      - oidc
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"